Move re-login message modal files into Home component folder
